Clean-up and refactor
