+++ .env
@@ -1,8 +1,8 @@
-VITE_FIREBASE_API_KEY="AIzaSyBVZKxviTn5mbnZnZQG2NHaR8HMnQledBw"
-VITE_FIREBASE_AUTH_DOMAIN="sodc-web.firebaseapp.com"
-VITE_FIREBASE_PROJECT_ID="sodc-web"
-VITE_FIREBASE_STORAGE_BUCKET="sodc-web.firebasestorage.app"
-VITE_FIREBASE_MESSAGING_SENDER_ID="931675247602"
-VITE_FIREBASE_APP_ID="1:931675247602:web:ff9805ec63d708f444516b"
-VITE_FIREBASE_MEASUREMENT_ID="G-J86JJGNY7S"
+VITE_FIREBASE_API_KEY=
+VITE_FIREBASE_AUTH_DOMAIN=
+VITE_FIREBASE_PROJECT_ID=
+VITE_FIREBASE_STORAGE_BUCKET=
+VITE_FIREBASE_MESSAGING_SENDER_ID=
+VITE_FIREBASE_APP_ID=
+VITE_FIREBASE_MEASUREMENT_ID=
