+++ src/constants/routes.ts
@@ -1,0 +1,20 @@
+/**
+ * Application route definitions
+ */
+
+export const ROUTES = {
+  HOME: "home",
+  ACCOUNT: "account",
+  PROFILE: "profile",
+  PERMISSIONS: "permissions",
+  MANAGE_USERS: "manageUsers",
+  APPROVE_USERS: "approveUsers",
+  REGISTER: "register",
+  PROFILE_COMPLETION: "profileCompletion",
+} as const;
+
+/**
+ * Route type
+ */
+export type Route = typeof ROUTES[keyof typeof ROUTES];
+
