+++ src/features/admin/components/ApproveUsers.tsx
@@ -15,15 +15,15 @@ import {
 } from "@mui/material";
 import { Visibility, CheckCircle } from "@mui/icons-material";
 import { executeQuery } from "firebase/data-connect";
-import { dataConnect } from "../config/firebase";
+import { dataConnect } from "../../../config/firebase";
 import { listUsersRef } from "@dataconnect/generated";
-import { colors } from "../config/colors";
-import PageHeader from "./PageHeader";
-import EditUserDialog from "./EditUserDialog";
-import { MEMBERSHIP_STATUS_OPTIONS } from "../constants";
+import { colors } from "../../../config/colors";
+import PageHeader from "../../../shared/components/PageHeader";
+import EditUserDialog from "../../profile/components/EditUserDialog";
+import { MEMBERSHIP_STATUS_OPTIONS } from "../../../constants";
 import { MembershipStatus } from "@dataconnect/generated";
-import type { SearchUser, UserData, PendingUser } from "../types";
-import { updateMembershipStatus } from "../shared/utils/firebaseFunctions";
+import type { SearchUser, PendingUser } from "../../../types";
+import { updateMembershipStatus } from "../../../shared/utils/firebaseFunctions";
 import { Snackbar } from "@mui/material";
 import "./ApproveUsers.css";
 
@@ -47,8 +47,8 @@ export default function ApproveUsers({ onBack }: ApproveUsersProps) {
     try {
       const ref = listUsersRef(dataConnect);
       const result = await executeQuery(ref);
-      if (result.error) {
-        throw new Error(result.error.message || "Failed to fetch users");
+      if (!result.data) {
+        throw new Error("Failed to fetch users");
       }
       // Filter for users with PENDING status
       const pending = (result.data?.users || []).filter(
