+++ src/features/auth/components/AuthGate.tsx
@@ -10,13 +10,13 @@ import {
   Link,
 } from "@mui/material";
 import { onAuthStateChanged, signInWithEmailAndPassword, signOut, type User } from "firebase/auth";
-import { auth } from "../config/firebase";
-import { useEnabledClaim } from "../hooks/useEnabledClaim";
-import AccountStatusMessage from "./AccountStatusMessage";
-import type { UserData } from "../types";
+import { auth } from "../../../config/firebase";
+import { useEnabledClaim } from "../../users/hooks/useEnabledClaim";
+import AccountStatusMessage from "../../users/components/AccountStatusMessage";
+import type { UserData } from "../../../types";
 import Register from "./Register";
 import EmailVerificationMessage from "./EmailVerificationMessage";
-import { colors } from "../config/colors";
+import { colors } from "../../../config/colors";
 
 interface AuthGateProps {
   userData?: UserData | null;
@@ -99,7 +99,7 @@ export default function AuthGate({ userData, onBack, onRegisterComplete, onProfi
 
     // If user is signed in but not enabled, show account status message
     if (!isEnabled) {
-      return <AccountStatusMessage userData={userData} onBack={onBack} />;
+      return <AccountStatusMessage userData={userData ?? null} onBack={onBack} />;
     }
 
     // User is signed in and enabled
