+++ src/features/users/components/AccountStatusMessage.tsx
@@ -7,9 +7,9 @@ import {
   Typography,
 } from "@mui/material";
 import { signOut } from "firebase/auth";
-import { auth } from "../config/firebase";
-import { colors } from "../config/colors";
-import type { UserData } from "../types";
+import { auth } from "../../../config/firebase";
+import { colors } from "../../../config/colors";
+import type { UserData } from "../../../types";
 
 interface AccountStatusMessageProps {
   userData: UserData | null;
