+++ src/features/users/components/UsersTable.tsx
@@ -13,8 +13,8 @@ import {
   Tooltip,
 } from "@mui/material";
 import { Edit, CheckCircle } from "@mui/icons-material";
-import type { SearchUser } from "../types";
-import { colors } from "../config/colors";
+import type { SearchUser } from "../../../types";
+import { colors } from "../../../config/colors";
 
 interface UsersTableProps {
   users: SearchUser[];
