+++ src/shared/components/SearchBar.css
