+++ src/shared/components/SearchBar.tsx
