+++ src/types/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Centralized type exports
+ * Import types from here for consistency
+ */
+
+export * from "./user";
+export * from "./auth";
+export * from "./api";
+
