--- src/utils/grantAdmin.ts
@@ -1,37 +1,0 @@
-import { getFunctions, httpsCallable } from "firebase/functions";
-import { firebaseApp } from "../config/firebase";
-
-interface GrantAdminRequest {
-  uid: string;
-}
-
-interface GrantAdminResponse {
-  success: boolean;
-  message?: string;
-}
-
-/**
- * Grants admin claim to a user
- * @param uid - The UID of the user to grant admin to
- * @returns Promise with the result
- */
-export async function grantAdminClaim(
-  uid: string
-): Promise<{ success: boolean; message?: string; error?: string }> {
-  try {
-    const functions = getFunctions(firebaseApp, "europe-west2");
-    const grantAdminCallable = httpsCallable<GrantAdminRequest, GrantAdminResponse>(
-      functions,
-      "grantAdmin"
-    );
-    
-    const result = await grantAdminCallable({ uid });
-    return { success: true, message: result.data.message };
-  } catch (error: any) {
-    return { 
-      success: false, 
-      error: error?.message || "Failed to grant admin claim" 
-    };
-  }
-}
-
