--- src/utils/updateMembershipStatus.ts
@@ -1,42 +1,0 @@
-import { getFunctions, httpsCallable } from "firebase/functions";
-import { firebaseApp } from "../config/firebase";
-import type { MembershipStatus } from "../dataconnect-generated";
-
-interface UpdateMembershipStatusRequest {
-  userId: string;
-  newStatus: MembershipStatus;
-}
-
-interface UpdateMembershipStatusResponse {
-  success: boolean;
-  message?: string;
-}
-
-/**
- * Updates membership status with server-side validation and enforcement
- * This function validates AND updates the membership status, preventing bypass
- * @param userId - The user ID whose status is being changed
- * @param newStatus - The desired new membership status
- * @returns Promise with the result
- */
-export async function updateMembershipStatus(
-  userId: string,
-  newStatus: MembershipStatus
-): Promise<{ success: boolean; error?: string }> {
-  try {
-    const functions = getFunctions(firebaseApp, "europe-west2");
-    const updateMembershipStatusCallable = httpsCallable<
-      UpdateMembershipStatusRequest,
-      UpdateMembershipStatusResponse
-    >(functions, "updateMembershipStatus");
-    
-    const result = await updateMembershipStatusCallable({ userId, newStatus });
-    return { success: result.data.success };
-  } catch (error: any) {
-    return { 
-      success: false, 
-      error: error?.message || "Failed to update membership status" 
-    };
-  }
-}
-
